Minor edits to editing.md (#79)
* Just fixing some typos

Just fixed a few typos. No major changes

* Added a section on the label bar

Added a short section describing the label bar. Feel free to not approve, edit, or move this section

* Added label bar image

Image of label bar from QGIS 3.10

* Added line break after label bar image

* Added toggle display of unplaced labels 

added toggle display of unplace labels to the label bar section

* Fixed some Typos

Fixed some typos and added an image of the magnet button referred to in the text

* added snapping magnet image